+++ Makefile
@@ -6,7 +6,7 @@
 #    By: kanghyki <kanghyki@student.42seoul.kr>     +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/06/01 00:54:06 by kanghyki          #+#    #+#              #
-#    Updated: 2022/06/12 04:39:10 by kanghyki         ###   ########.fr        #
+#    Updated: 2022/06/12 17:09:43 by kanghyki         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -31,6 +31,12 @@ PARSER_SRC		=	parser.c\
 					parser_common_utils.c\
 					parser_free_utils.c\
 					parser_additional.c
+#--------------[ ENV ]----------------
+ENV_DIR			=	src/environment
+ENV_SRC			=	env.c\
+					env_utils.c\
+					env_free_utils.c\
+					env_execute.c
 #--------------[ EXCUSE ]----------------
 EXCUSE_DIR		=	src/excuse_cmd
 EXCUSE_SRC		=	#cmd.c
@@ -38,14 +44,13 @@ EXCUSE_SRC		=	#cmd.c
 NAME			=	minishell	
 OBJ_DIR			=	objects
 SRC_DIR			=	src
-SRC				=	main.c\
-					env_utils.c\
-					env_execute.c
+SRC				=	main.c
 
 OBJS			=	$(addprefix $(OBJ_DIR)/, $(SRC:.c=.o))\
 					$(addprefix $(OBJ_DIR)/, $(PARSER_SRC:.c=.o))\
 					$(addprefix $(OBJ_DIR)/, $(TOKEN_SRC:.c=.o))\
-					$(addprefix $(OBJ_DIR)/, $(EXCUSE_SRC:.c=.o))
+					$(addprefix $(OBJ_DIR)/, $(ENV_SRC:.c=.o))
+					#$(addprefix $(OBJ_DIR)/, $(EXCUSE_SRC:.c=.o))
 #-----------------[ CMD ]-------------------
 CC				=	gcc
 CFLAGS			=	-g # -Wall -Wextra -Werror
@@ -59,7 +64,7 @@ LDLIBS			=	-l ft -L $(LIBFT_DIR)\
 AR				=	ar -rcs
 RM				=	rm -rf
 
-vpath %.c $(SRC_DIR) $(TOKEN_DIR) $(PARSER_DIR) $(EXCUSE_DIR)
+vpath %.c $(SRC_DIR) $(TOKEN_DIR) $(PARSER_DIR) $(ENV_DIR) #$(EXCUSE_DIR)
 
 all: $(NAME)
