+++ src/environment/env_execute.c
@@ -6,11 +6,11 @@
 /*   By: kanghyki <kanghyki@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/10 14:19:31 by kanghyki          #+#    #+#             */
-/*   Updated: 2022/06/12 03:46:47 by kanghyki         ###   ########.fr       */
+/*   Updated: 2022/06/12 17:00:11 by kanghyki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/minishell.h"
+#include "../../include/minishell.h"
 
 int	ft_strcmp(const char *s1, const char *s2)
 {
@@ -28,29 +28,23 @@ int	ft_strcmp(const char *s1, const char *s2)
 
 void	ft_execute_export(t_argument *arg)
 {
-	const int	COMMAND_ARG_POSITION = 1;
-
-	if (arg->pa_argument[COMMAND_ARG_POSITION] != 0)
-		*(arg->env) = ft_set_env(*(arg->env), arg->pa_argument[COMMAND_ARG_POSITION]);
+	if (arg->pa_argument[1] != 0)
+		*(arg->env) = ft_set_env(*(arg->env), arg->pa_argument[1]);
 	else
 		ft_print_env(*(arg->env));
 }
 
 void	ft_execute_env(t_argument *arg)
 {
-	const int	COMMAND_ARG_POSITION = 1;
-
-	if (arg->pa_argument[COMMAND_ARG_POSITION] != 0)
-		*(arg->env) = ft_set_env(*(arg->env), arg->pa_argument[COMMAND_ARG_POSITION]);
+	if (arg->pa_argument[1] != 0)
+		*(arg->env) = ft_set_env(*(arg->env), arg->pa_argument[1]);
 	ft_print_env(*(arg->env));
 }
 
 void	ft_execute_unset(t_argument *arg)
 {
-	const int	COMMAND_ARG_POSITION = 1;
-
-	if (arg->pa_argument[COMMAND_ARG_POSITION] != 0)
-		*(arg->env) = ft_unset_env(*(arg->env), arg->pa_argument[COMMAND_ARG_POSITION]);
+	if (arg->pa_argument[1] != 0)
+		*(arg->env) = ft_unset_env(*(arg->env), arg->pa_argument[1]);
 }
 
 void	ft_env_simple_command_test(t_argument *arg)
